update login và add room

+++ app/src/main/java/com/example/dormitorymanager/Model/Rooms.kt
@@ -3,9 +3,9 @@ package com.example.dormitorymanager.Model
 data class Rooms(var _id:String,var beds: String,
                 var description:String, var location: String
                 ,var name:String,
-                 var prices:Double,
+                 var prices:Long,
                  var status: String) {
 
     // Hàm khởi tạo không đối số
-    constructor() : this("", "", "8", "1000000.0", "", 100000.0, "")
+    constructor() : this("", "", "8", "1000000.0", "", 100000, "")
 }
